code/lesson6/requirements.txt
langchain==1.1.0
langchain-openai==1.1.0
langchain-community==0.4.1
python-dotenv==1.2.1